test: extract shared helpers, split monolith, drop tautological tests
The 3978-line `AIChat_Watch_AppTests.swift` was a single-class catch-all
covering 100 mostly-unrelated tests, with private helpers and the same
fixture constants duplicated across siblings. UI tests on Watch and iOS
each carried their own copy of `attachScreenshot` / `attachDebugHierarchy`
/ `waitForHittable`. Sanity audit also flagged a few low-value tests.

This refactor:

- Adds `AIChat Watch AppTests/TestSupport.swift` with shared fixtures
  (`onePixelPNGBase64`, `makeOnePixelImageAttachment`,
  `makeTemporaryRootURL`, `makeDirectModeAppConfiguration`).
  `ConversationRepositoryTests` and `ICloudConversationSyncServiceTests`
  now use them instead of redefining the same constants/helpers.

- Adds `UITestSupport.swift` to both UI test targets via an `XCTestCase`
  extension. Strips duplicated screenshot/hierarchy/hittability helpers
  from `AIChat_Watch_AppUITests.swift` (-67 lines) and
  `AIChat_iOS_AppUITests.swift` (-59 lines).

- Splits the monolith into six focused subject files (27 tests moved):
  - `PureFunctionUtilityTests.swift` — title, deeplink, completion feedback
  - `ConversationSortingAndComposerTests.swift` — recency sort, draft composer
  - `ConversationHistoryRenderBudgetTests.swift` — render budget rules
  - `AIModelCatalogAndCompletionErrorTests.swift` — model catalog + completion error
  - `PromptAssemblyTests.swift` — `AIContextAssembler` + preset library
  - `RelayRequestShapeTests.swift` — `RelayAIClient.makeRelayRequest`

- Merges two trivially tautological tests in
  `ConversationAutoScrollControllerTests` into a single test that pins
  the same invariant in one place.

- Replaces five copy-pasted persistence dances in `SettingsServiceTests`
  with a single `assertPersistsAcrossRestart` helper; trims the file
  from 231 to 197 lines without changing coverage.

Behaviour is unchanged. Test target uses
`PBXFileSystemSynchronizedRootGroup`, so the new files are picked up
automatically with no `project.pbxproj` edit.
