+++ AIChat Watch AppTests/AIChat_Watch_AppTests.swift
@@ -15,47 +15,6 @@ final class AIChat_Watch_AppTests: XCTestCase {
     @MainActor
     private static var retainedConversationListPerformanceStores: [ChatStore] = []
 
-    private let onePixelPNGBase64 =
-        "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mP8/x8AAwMCAO+aW2QAAAAASUVORK5CYII="
-
-    func testSuggestedTitleUsesCollapsedWhitespaceAndTruncates() {
-        let input = "  Build   a production ready Apple Watch assistant with photo upload   "
-        let title = ConversationThread.suggestedTitle(from: input)
-
-        XCTAssertEqual(title, "Build a production ready A...")
-    }
-
-    func testAIChatDeepLinkParsesActivationImport() {
-        let url = URL(string: "aichat://activation/import?code=abcd-1234-efgh")!
-
-        XCTAssertEqual(
-            AIChatDeepLink(url),
-            .activationImport("ABCD-1234-EFGH")
-        )
-    }
-
-    func testAIChatDeepLinkParsesNewConversation() {
-        let url = URL(string: "aichat://conversation/new")!
-
-        XCTAssertEqual(AIChatDeepLink(url), .newConversation)
-    }
-
-    func testCompletionFeedbackForegroundPresentationOptionsEnableSound() {
-        let options = CompletionFeedbackEvent.foregroundPresentationOptions(
-            forNotificationIdentifier: CompletionFeedbackEvent.transcriptionCompleted.notificationIdentifier
-        )
-
-        XCTAssertTrue(options.contains(.sound))
-    }
-
-    func testCompletionFeedbackForegroundPresentationOptionsIgnoreUnknownNotifications() {
-        let options = CompletionFeedbackEvent.foregroundPresentationOptions(
-            forNotificationIdentifier: "some-other-notification"
-        )
-
-        XCTAssertEqual(options, [])
-    }
-
     func testContextWindowMapsRolesAndPreservesLatestMessages() {
         let imageData = Data([0x01, 0x02, 0x03])
         let attachment = ChatImageAttachment(
@@ -411,156 +370,6 @@ final class AIChat_Watch_AppTests: XCTestCase {
         XCTAssertNil(requestBody.generationConfig.mediaResolution)
         XCTAssertEqual(requestBody.generationConfig.maxOutputTokens, 65_536)
         XCTAssertEqual(requestBody.safetySettings, GeminiSafetySetting.aiStudioDefaults)
-    }
-
-    func testDefaultPromptModeOmitsSystemInstruction() {
-        let conversation = ConversationThread(
-            messages: [ChatMessage(role: .user, text: "Summarize this thread")],
-            aiConfiguration: ConversationAIConfiguration(
-                model: "gemini-3.1-pro-preview",
-                thinkingIntensity: .balanced,
-                systemPromptMode: .default
-            )
-        )
-
-        let client = GeminiAPIClient(
-            configuration: AppConfiguration(
-                backendMode: .direct,
-                geminiAPIKey: "test",
-                geminiModel: "gemini-3.1-pro-preview",
-                geminiTranscriptionModel: "gemini-3-flash-preview",
-                relayBaseURL: nil,
-                relayBearerToken: nil,
-                relayStreamPath: "v1/chat/stream",
-                appGroupIdentifier: nil
-            )
-        )
-
-        let requestBody = client.makeRequestBody(for: conversation)
-
-        XCTAssertNil(requestBody.systemInstruction)
-    }
-
-    func testContextAssemblerOmitsCasualFocusWithoutOpenLoops() {
-        let conversation = ConversationThread(
-            title: "Dinner",
-            messages: [
-                ChatMessage(role: .user, text: "今晚吃什么好？")
-            ],
-            focusState: ConversationFocusState(
-                kind: .casual,
-                title: "Dinner chat",
-                focusNote: "User is casually asking for dinner ideas.",
-                openLoops: []
-            )
-        )
-
-        let assembled = AIContextAssembler.assembleReplyContext(
-            for: conversation,
-            configuration: ConversationAIConfiguration(model: "gemini-3-flash-preview")
-        )
-
-        XCTAssertEqual(assembled.mode, .casual)
-        XCTAssertEqual(assembled.systemInstructionParts?.count, 1)
-        XCTAssertEqual(assembled.systemInstructionParts?.first?.text, AIContextAssembler.conciseSystemPrompt)
-    }
-
-    func testContextAssemblerIncludesTeachingArtifactsInSystemInstructionParts() {
-        let conversation = ConversationThread(
-            title: "函数导数",
-            messages: [
-                ChatMessage(role: .user, text: "已知 f(x)=x^2+2x，帮我讲一下导数思路")
-            ],
-            focusState: ConversationFocusState(
-                kind: .teaching,
-                title: "导数讲解",
-                focusNote: "围绕二次函数求导，已经确认要先回顾导数定义，再带学生做一步步推导。",
-                openLoops: ["还要解释为什么常数项求导为 0"]
-            ),
-            memoryItems: [
-                ConversationMemoryItem(
-                    text: "学生更适合先给直观思路，再给公式化写法。",
-                    keywords: ["思路", "公式"]
-                )
-            ],
-            pinnedMemories: [
-                PinnedMemoryItem(
-                    text: "用户是高中理科中文教学场景。",
-                    keywords: ["高中", "理科", "教学"],
-                    scope: .conversation
-                )
-            ],
-            archiveSegments: [
-                ConversationArchiveSegment(
-                    title: "上一次函数题",
-                    summary: "之前已经讲过一次单调性判断，学生在符号书写上容易漏掉定义域。",
-                    keywords: ["函数", "定义域"],
-                    openLoops: ["本轮还可以顺手复习定义域"],
-                    sourceMessageIDs: []
-                )
-            ]
-        )
-
-        let assembled = AIContextAssembler.assembleReplyContext(
-            for: conversation,
-            configuration: ConversationAIConfiguration(model: "gemini-3-flash-preview")
-        )
-        let instructionText = assembled.systemInstructionParts?.compactMap(\.text).joined(separator: "\n\n")
-
-        XCTAssertEqual(assembled.mode, .teaching)
-        XCTAssertTrue(instructionText?.contains("Current focus:") == true)
-        XCTAssertTrue(instructionText?.contains("Pinned memory:") == true)
-        XCTAssertTrue(instructionText?.contains("Relevant conversation memory:") == true)
-        XCTAssertTrue(instructionText?.contains("Archived context:") == true)
-    }
-
-    func testCustomSystemPromptOverridesBuiltInPrompt() {
-        let conversation = ConversationThread(
-            messages: [ChatMessage(role: .user, text: "Summarize this thread")],
-            aiConfiguration: ConversationAIConfiguration(
-                model: "gemini-3-flash-preview",
-                customSystemPrompt: "You are a strict proofreader. Return only corrected text."
-            )
-        )
-
-        let client = GeminiAPIClient(
-            configuration: AppConfiguration(
-                backendMode: .direct,
-                geminiAPIKey: "test",
-                geminiModel: "gemini-3-flash-preview",
-                geminiTranscriptionModel: "gemini-3-flash-preview",
-                relayBaseURL: nil,
-                relayBearerToken: nil,
-                relayStreamPath: "v1/chat/stream",
-                appGroupIdentifier: nil
-            )
-        )
-
-        let requestBody = client.makeRequestBody(for: conversation)
-
-        XCTAssertEqual(
-            requestBody.systemInstruction?.parts.first?.text,
-            "You are a strict proofreader. Return only corrected text."
-        )
-    }
-
-    func testPromptPresetLibraryIncludesBuiltInsByDefault() {
-        let presets = PromptPreset.resolvedLibrary(from: [])
-
-        XCTAssertTrue(
-            presets.contains(where: {
-                $0.id == PromptPreset.builtInConversationID &&
-                $0.kind == .conversation &&
-                $0.isBuiltIn
-            })
-        )
-        XCTAssertTrue(
-            presets.contains(where: {
-                $0.id == PromptPreset.builtInTranscriptionID &&
-                $0.kind == .transcription &&
-                $0.isBuiltIn
-            })
-        )
     }
 
     func testGemini25RequestUsesThinkingBudget() {
@@ -1076,113 +885,6 @@ final class AIChat_Watch_AppTests: XCTestCase {
         XCTAssertEqual(currentText, "Answer ")
     }
 
-    func testRelayRequestCarriesThinkingOutputTokenBudget() {
-        let conversation = ConversationThread(
-            messages: [ChatMessage(role: .user, text: "Write a deeper analysis")],
-            aiConfiguration: ConversationAIConfiguration(
-                model: "gemini-3-flash-preview",
-                thinkingIntensity: .deep
-            )
-        )
-
-        let client = RelayAIClient(
-            configuration: AppConfiguration(
-                backendMode: .relay,
-                geminiAPIKey: nil,
-                geminiModel: "gemini-3-flash-preview",
-                geminiTranscriptionModel: "gemini-3-flash-preview",
-                relayBaseURL: URL(string: "http://127.0.0.1:8787"),
-                relayBearerToken: "token",
-                relayStreamPath: "v1/chat/stream",
-                appGroupIdentifier: nil
-            )
-        )
-
-        let request = client.makeRelayRequest(for: conversation)
-
-        XCTAssertEqual(request.maxOutputTokens, 65_536)
-        XCTAssertNil(request.systemPrompt)
-        XCTAssertEqual(request.systemInstructionParts?.first?.text, AIContextAssembler.conciseSystemPrompt)
-    }
-
-    func testRelayRequestCarriesGeminiToolFlags() {
-        let conversation = ConversationThread(
-            messages: [ChatMessage(role: .user, text: "Search the web and run a quick calculation.")],
-            aiConfiguration: ConversationAIConfiguration(
-                model: "gemini-3-flash-preview",
-                usesGoogleSearch: true,
-                usesCodeExecution: true
-            )
-        )
-
-        let client = RelayAIClient(
-            configuration: AppConfiguration(
-                backendMode: .relay,
-                geminiAPIKey: nil,
-                geminiModel: "gemini-3-flash-preview",
-                geminiTranscriptionModel: "gemini-3-flash-preview",
-                relayBaseURL: URL(string: "http://127.0.0.1:8787"),
-                relayBearerToken: "token",
-                relayStreamPath: "v1/chat/stream",
-                appGroupIdentifier: nil
-            )
-        )
-
-        let request = client.makeRelayRequest(for: conversation)
-
-        XCTAssertTrue(request.usesGoogleSearch)
-        XCTAssertTrue(request.usesCodeExecution)
-    }
-
-    func testRelayRequestCarriesStoredAssistantModelParts() {
-        let storedParts = [
-            GeminiPart(
-                text: "Intermediate reasoning",
-                inlineData: nil,
-                thought: true,
-                thoughtSignature: "sig-thought"
-            ),
-            GeminiPart(
-                text: "Final answer",
-                inlineData: nil,
-                thought: false,
-                thoughtSignature: "sig-answer"
-            )
-        ]
-        let conversation = ConversationThread(
-            messages: [
-                ChatMessage(role: .user, text: "Question"),
-                ChatMessage(
-                    role: .assistant,
-                    text: "Final answer",
-                    thoughtSummary: "Intermediate reasoning",
-                    modelResponseParts: storedParts
-                ),
-                ChatMessage(role: .user, text: "Follow up")
-            ],
-            aiConfiguration: ConversationAIConfiguration(model: "gemini-3-flash-preview")
-        )
-
-        let client = RelayAIClient(
-            configuration: AppConfiguration(
-                backendMode: .relay,
-                geminiAPIKey: nil,
-                geminiModel: "gemini-3-flash-preview",
-                geminiTranscriptionModel: "gemini-3-flash-preview",
-                relayBaseURL: URL(string: "http://127.0.0.1:8787"),
-                relayBearerToken: "token",
-                relayStreamPath: "v1/chat/stream",
-                appGroupIdentifier: nil
-            )
-        )
-
-        let request = client.makeRelayRequest(for: conversation)
-
-        XCTAssertEqual(request.messages.count, 3)
-        XCTAssertEqual(request.messages[1].role, "assistant")
-        XCTAssertEqual(request.messages[1].modelResponseParts, storedParts)
-    }
-
     func testMemoryExtractionIgnoresHiddenAssistantModelPartsWithoutVisibleText() async {
         let extractor = CapturingMemoryExtractionClient(
             response: ConversationMemoryExtractionResponse()
@@ -1317,49 +1019,6 @@ final class AIChat_Watch_AppTests: XCTestCase {
         XCTAssertTrue(request.prompt.contains("User: Use the Tokyo one and keep it short."))
         XCTAssertEqual(request.audio.mimeType, "audio/aac")
         XCTAssertEqual(request.audio.base64Data, audioData.base64EncodedString())
-    }
-
-    func testModelCatalogUsesDesktopScaleOutputBudgetForSupportedGeminiModels() {
-        XCTAssertEqual(AIModelCatalog.maxOutputTokens(for: "gemini-3-flash-preview"), 65_536)
-        XCTAssertEqual(AIModelCatalog.maxOutputTokens(for: "gemini-3.1-pro-preview"), 65_536)
-        XCTAssertEqual(AIModelCatalog.maxOutputTokens(for: "gemini-2.5-flash"), 65_536)
-        XCTAssertEqual(AIModelCatalog.maxOutputTokens(for: "custom-model"), 8_192)
-    }
-
-    func testModelCatalogLimitsExtremeThinkingToGemini31Pro() {
-        XCTAssertEqual(
-            AIModelCatalog.availableThinkingIntensities(for: "gemini-3.1-pro-preview"),
-            [.fast, .balanced, .deep, .extreme]
-        )
-        XCTAssertEqual(
-            AIModelCatalog.availableThinkingIntensities(for: "gemini-3-flash-preview"),
-            [.fast, .balanced, .deep]
-        )
-        XCTAssertEqual(
-            AIModelCatalog.normalizedThinkingIntensity(.extreme, for: "gemini-3-flash-preview"),
-            .deep
-        )
-    }
-
-    func testGeminiCompletionErrorRequiresTerminalFinishReason() {
-        XCTAssertEqual(geminiCompletionError(for: nil), .incompleteResponse)
-        XCTAssertEqual(geminiCompletionError(for: "STOP"), nil)
-        XCTAssertEqual(geminiCompletionError(for: "MAX_TOKENS"), .truncated)
-    }
-
-    func testRelayCompletionErrorRequiresDoneEventAndStopFinishReason() {
-        XCTAssertEqual(
-            relayCompletionError(didReceiveDoneEvent: false, finishReason: "STOP"),
-            .incompleteResponse
-        )
-        XCTAssertEqual(
-            relayCompletionError(didReceiveDoneEvent: true, finishReason: "STOP"),
-            nil
-        )
-        XCTAssertEqual(
-            relayCompletionError(didReceiveDoneEvent: true, finishReason: "MAX_TOKENS"),
-            .truncated
-        )
     }
 
     func testTranscriptionRequestUsesContextAndAudio() throws {
@@ -3267,203 +2926,6 @@ final class AIChat_Watch_AppTests: XCTestCase {
         )
         XCTAssertEqual(persistedConversation.messages.count, 1)
         XCTAssertEqual(persistedConversation.messages.first?.role, .user)
-    }
-
-    func testConversationRecencySortPrefersNewestUpdatedAtThenCreatedAt() {
-        let older = ConversationThread(
-            id: UUID(uuidString: "00000000-0000-0000-0000-000000000001")!,
-            title: "older",
-            createdAt: Date(timeIntervalSince1970: 100),
-            updatedAt: Date(timeIntervalSince1970: 200)
-        )
-        let newer = ConversationThread(
-            id: UUID(uuidString: "00000000-0000-0000-0000-000000000002")!,
-            title: "newer",
-            createdAt: Date(timeIntervalSince1970: 150),
-            updatedAt: Date(timeIntervalSince1970: 300)
-        )
-        let tiedUpdateNewerCreate = ConversationThread(
-            id: UUID(uuidString: "00000000-0000-0000-0000-000000000003")!,
-            title: "tie",
-            createdAt: Date(timeIntervalSince1970: 180),
-            updatedAt: Date(timeIntervalSince1970: 300)
-        )
-
-        let sorted = [older, newer, tiedUpdateNewerCreate]
-            .sorted(by: ConversationThread.sortsByMostRecentFirst)
-
-        XCTAssertEqual(sorted.map(\.title), ["tie", "newer", "older"])
-    }
-
-    func testDraftTextComposerAppendsChineseTranscriptWithPunctuationAwareSeparator() {
-        let merged = DraftTextComposer.appended(
-            existing: "帮我写一封邮件给张三",
-            addition: "内容是明天下午三点开会。"
-        )
-
-        XCTAssertEqual(merged, "帮我写一封邮件给张三，内容是明天下午三点开会。")
-    }
-
-    func testDraftTextComposerKeepsEnglishContinuationInline() {
-        let merged = DraftTextComposer.appended(
-            existing: "Summarize the meeting",
-            addition: "and call out the blockers."
-        )
-
-        XCTAssertEqual(merged, "Summarize the meeting and call out the blockers.")
-    }
-
-    func testConversationHistoryRenderBudgetDefersSmallHeavyConversation() {
-        let messages = makeMessages(
-            count: 8,
-            text: String(repeating: "Large content block ", count: 280)
-        )
-
-        XCTAssertTrue(
-            ConversationHistoryRenderBudget.shouldDeferInitialRendering(
-                in: messages,
-                threshold: 48
-            )
-        )
-        XCTAssertEqual(
-            ConversationHistoryRenderBudget.visibleMessageCount(
-                in: messages,
-                budget: 10
-            ),
-            2
-        )
-    }
-
-    func testConversationHistoryRenderBudgetKeepsSmallLightConversationEager() {
-        let messages = makeMessages(
-            count: 8,
-            text: "Short reply"
-        )
-
-        XCTAssertFalse(
-            ConversationHistoryRenderBudget.shouldDeferInitialRendering(
-                in: messages,
-                threshold: 48
-            )
-        )
-        XCTAssertEqual(
-            ConversationHistoryRenderBudget.visibleMessageCount(
-                in: messages,
-                budget: 10
-            ),
-            messages.count
-        )
-    }
-
-    func testConversationHistoryRenderBudgetAlwaysKeepsNewestMessageVisible() {
-        let messages = makeMessages(
-            count: 1,
-            text: String(repeating: "Very long answer ", count: 400)
-        )
-
-        XCTAssertEqual(
-            ConversationHistoryRenderBudget.visibleMessageCount(
-                in: messages,
-                budget: 1
-            ),
-            1
-        )
-    }
-
-    func testConversationHistoryRenderBudgetAlwaysKeepsLatestExchangeVisible() {
-        let messages = [
-            ChatMessage(role: .user, text: "Older question"),
-            ChatMessage(role: .assistant, text: "Older answer"),
-            ChatMessage(role: .user, text: "Latest question"),
-            ChatMessage(role: .assistant, text: String(repeating: "Very long answer ", count: 400))
-        ]
-
-        XCTAssertEqual(
-            ConversationHistoryRenderBudget.visibleMessageCount(
-                in: messages,
-                budget: 1
-            ),
-            2
-        )
-    }
-
-    func testConversationHistoryRenderBudgetKeepsPendingLatestUserMessageVisible() {
-        let messages = [
-            ChatMessage(role: .user, text: "Older question"),
-            ChatMessage(role: .assistant, text: "Older answer"),
-            ChatMessage(role: .user, text: String(repeating: "Latest pending request ", count: 300))
-        ]
-
-        XCTAssertEqual(
-            ConversationHistoryRenderBudget.visibleMessageCount(
-                in: messages,
-                budget: 1
-            ),
-            1
-        )
-    }
-
-    func testConversationHistoryRenderBudgetKeepsLoadMoreAnchorAtLastHiddenMessage() {
-        let messages = makeMessages(
-            count: 5,
-            text: "Short reply"
-        )
-
-        XCTAssertEqual(
-            ConversationHistoryRenderBudget.lastHiddenMessageID(
-                in: messages,
-                budget: 2
-            ),
-            messages[2].id
-        )
-    }
-
-    func testConversationHistoryRenderBudgetLoadMoreAlwaysRevealsAnotherMessage() {
-        let messages = [
-            ChatMessage(role: .user, text: "Newest question"),
-            ChatMessage(
-                role: .assistant,
-                text: Array(
-                    repeating: """
-                    ## 公式
-                    $$\\int_0^1 x^4 dx = \\frac{1}{5}$$
-                    """,
-                    count: 120
-                ).joined(separator: "\n")
-            ),
-            ChatMessage(role: .user, text: "Latest follow-up")
-        ]
-
-        let currentBudget = 1
-        let nextBudget = ConversationHistoryRenderBudget.budgetForLoadingOlderMessages(
-            in: messages,
-            currentBudget: currentBudget,
-            preferredIncrement: 1
-        )
-
-        XCTAssertEqual(
-            ConversationHistoryRenderBudget.visibleMessageCount(
-                in: messages,
-                budget: currentBudget
-            ),
-            1
-        )
-        XCTAssertGreaterThanOrEqual(
-            ConversationHistoryRenderBudget.visibleMessageCount(
-                in: messages,
-                budget: nextBudget
-            ),
-            2
-        )
-    }
-
-    private func makeMessages(count: Int, text: String) -> [ChatMessage] {
-        (0..<count).map { index in
-            ChatMessage(
-                role: index.isMultiple(of: 2) ? .user : .assistant,
-                text: text
-            )
-        }
     }
 
     @MainActor
